Updating the Post form with quantity
Updating the Post Form including quantity, by creating a input field in the PostCreation.tsx with a label of Quantity, and get the quantity in the backend main class CreatePost class that exists in Main\create_post.php, then in the object in create_post.php. There is also an modification in get_donations.php.